refactor: package compvision.lineutils
